Version ..,  Revision 22023,  Hash 8332aa902
MODIFIED *  .gitattributes config/configure.ac config/mk/local/makefile sbin/compilation/object_remove.sh sbin/compilation/object_save_and_restore.sh sbin/yambo_versions_update.tcsh

DELETED *  include/version.inc

RENAMED *  include/driver/version.h -> include/version/version.h config/version/version.m4 -> include/version/version.m4

Bugs:
- Fixed bug spotted by DS (see #718).

Changes:
- Moved all version-related files in include/version preparing the simplification of the update procedure.

Patch sent by:  Andrea Marini <andrea.marini@cnr.it>

+++ sbin/compilation/object_save_and_restore.sh
@@ -47,6 +47,7 @@ count_mod_res=`ls -1 $dir/$restore_dir/*.mod 2>/dev/null | wc -l`
 count_f90_res=`ls -1 $dir/$restore_dir/*.f90 2>/dev/null | wc -l`
 #
 # Save files
+#
 if [ $count_obj != 0 ]; then
  if [ ! -d $dir/$save_dir ] ; then
   source $compdir/sbin/compilation/verbosity.sh "object_save_and_restore.sh: mkdir $save_dir"
@@ -127,14 +128,16 @@ if [[ -d $dir/$restore_dir ]] && [[ ! "$restore_dir" == "$save_dir" ]] ; then
   done
  fi
  cd $path_back
- if [[ -f $dir/$restore_dir/$library ]] || [  ! "$library" == "NONE" ]; then
+ if [[ -f $dir/$restore_dir/$library ]] && [[ ! "$library" == "NONE" ]]; then
   source $compdir/sbin/compilation/verbosity.sh "object_save_and_restore.sh: removing & linking $compdir/$dir/$restore_dir/$library in $compdir/lib"
   cd $compdir/lib
   rm -f $library
   ln -s $compdir/$dir/$restore_dir/$library ./
+  DIR_restored="yes"
+ else
+  DIR_restored="no"
  fi
  cd $path_back
- DIR_restored="yes"
  source $compdir/sbin/compilation/fix_locks.sh
 fi
